Clean and bugfixes

+++ p3_a_tokenize_datasets.py
@@ -2,7 +2,7 @@ import numpy as np
 from tokenizers import Tokenizer
 from datasets import load_from_disk
 import random
-
+import os
 # ── config ────────────────────────────────────────────────────────
 DTYPE          = np.uint16   # alcanza para vocab ≤ 65535 (tu vocab=32000 ✓)
 # ─────────────────────────────────────────────────────────────────
@@ -13,56 +13,60 @@ def build_token_bin(root_dir,model_name,flush_every,log_every,datasets_versions)
     WIKI_EN_PATH   = root_dir + r"wiki_en_clean"
     OUT_BIN        = root_dir + model_name + r"_tokens.bin"
 
-    tokenizer = Tokenizer.from_file(TOKENIZER_PATH)
-    eos_id    = tokenizer.token_to_id("<eos>")
-
-    print("Cargando datasets...")
-    wiki_es = load_from_disk(WIKI_ES_PATH)
-    wiki_en = load_from_disk(WIKI_EN_PATH)
-
-    n_es = len(wiki_es)
-    n_en = len(wiki_en)
-
-    # índices: 0..n_es-1 son ES, n_es..n_es+n_en-1 son EN
-    all_indices = list(range(n_es + n_en))
-
     total_tokens = 0
-    buf = []
+    if not os.path.exists(OUT_BIN):
 
-    with open(OUT_BIN, "wb") as f:
-        for version in range(datasets_versions):
-            print(f"\n── Versión {version+1}/{datasets_versions} ──")
-            random.seed(version)
-            random.shuffle(all_indices)
+        tokenizer = Tokenizer.from_file(TOKENIZER_PATH)
+        eos_id    = tokenizer.token_to_id("<eos>")
 
-            for i, idx in enumerate(all_indices):
-                if idx < n_es:
-                    text = wiki_es[idx]["text"]
-                else:
-                    text = wiki_en[idx - n_es]["text"]
+        print("Cargando datasets...")
+        wiki_es = load_from_disk(WIKI_ES_PATH)
+        wiki_en = load_from_disk(WIKI_EN_PATH)
 
-                ids = tokenizer.encode(text).ids
-                buf.extend(ids)
-                buf.append(eos_id)
+        n_es = len(wiki_es)
+        n_en = len(wiki_en)
 
-                if len(buf) >= flush_every:
-                    arr = np.array(buf, dtype=DTYPE)
-                    f.write(arr.tobytes())
-                    total_tokens += len(buf)
-                    buf = []
+        # índices: 0..n_es-1 son ES, n_es..n_es+n_en-1 son EN
+        all_indices = list(range(n_es + n_en))
 
-                if i % log_every == 0:
-                    print(f"  art {i:,} | tokens escritos: {total_tokens:,}")
+        buf = []
 
-        if buf:
-            arr = np.array(buf, dtype=DTYPE)
-            f.write(arr.tobytes())
-            total_tokens += len(buf)
+        with open(OUT_BIN, "wb") as f:
+            for version in range(datasets_versions):
+                print(f"\n── Versión {version+1}/{datasets_versions} ──")
+                random.seed(version)
+                random.shuffle(all_indices)
 
-    size_gb = total_tokens * 2 / 1e9
-    print(f"\nListo: {OUT_BIN}")
-    print(f"Total tokens : {total_tokens:,} \n    ({total_tokens/datasets_versions} tokens por pasada)")
-    print(f"Tamaño disco : {size_gb:.2f} GB")
+                for i, idx in enumerate(all_indices):
+                    if idx < n_es:
+                        text = wiki_es[idx]["text"]
+                    else:
+                        text = wiki_en[idx - n_es]["text"]
+
+                    ids = tokenizer.encode(text).ids
+                    buf.extend(ids)
+                    buf.append(eos_id)
+
+                    if len(buf) >= flush_every:
+                        arr = np.array(buf, dtype=DTYPE)
+                        f.write(arr.tobytes())
+                        total_tokens += len(buf)
+                        buf = []
+
+                    if i % log_every == 0:
+                        print(f"  art {i:,} | tokens escritos: {total_tokens:,}")
+
+            if buf:
+                arr = np.array(buf, dtype=DTYPE)
+                f.write(arr.tobytes())
+                total_tokens += len(buf)
+
+        size_gb = total_tokens * 2 / 1e9
+        print(f"\nListo: {OUT_BIN}")
+        print(f"Total tokens : {total_tokens:,} \n    ({total_tokens/datasets_versions} tokens por pasada)")
+        print(f"Tamaño disco : {size_gb:.2f} GB")
+    else:
+        total_tokens = os.path.getsize(OUT_BIN)/2
     return total_tokens
 
 if __name__ == "__main__":
